backend/api/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
asyncpg==0.30.0
sqlmodel==0.0.24
Werkzeug==3.1.3
tqdm==4.67.1
aioboto3==15.0.0
aiohttp==3.11.7
slowapi==0.1.9
alembic==1.16.2
psycopg2-binary==2.9.10
python-dotenv>=1.0.1
openai==1.93.0
pydantic==2.11.7
requests==2.32.4
ujson==5.10.0
python-jose==3.3.0
passlib==1.7.4
httpx==0.28.1
APScheduler==3.11.0
aiofiles==24.1.0
aiocache==0.12.3
unidecode==1.3.8
python-multipart==0.0.20
pytz==2025.1
pandas==2.3.0
openpyxl==3.1.5
tabulate==0.9.0
posthog==6.0.4